+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [lib] `term_image.TermImageWarning`; pacage specific warning category ([#50]).
 - [lib] `term_image.set_query_timeout()` to set global query timeout ([3b658f3]).
 - [lib] Auto background color i.e using the terminal's default BG color for transparent images ([#54])
+- [lib] `ImageIterator.loop_no` property.
 - [cli] `--style` command-line option for render style selection ([#37]).
 - [cli] `kitty` render style choice for the `--style` command-line option ([#39]).
 - [cli] `--force-style` to bypass render style support checks ([#44]).